move spacefinder types to space-filler.ts
