+++ static/src/javascripts/projects/commercial/modules/article-body-adverts.ts
@@ -4,6 +4,11 @@ import { getBreakpoint, getViewport } from 'lib/detect-viewport';
 import config from '../../../lib/config';
 import { mediator } from '../../../lib/mediator';
 import { spaceFiller } from '../../common/modules/article/space-filler';
+import type {
+	SpacefinderItem,
+	SpacefinderRules,
+	SpacefinderWriter,
+} from '../../common/modules/article/space-filler';
 import { commercialFeatures } from '../../common/modules/commercial/commercial-features';
 import { initCarrot } from './carrot-traffic-driver';
 import { addSlot } from './dfp/add-slot';
