+++ static/src/javascripts/projects/commercial/modules/carrot-traffic-driver.ts
@@ -2,6 +2,10 @@ import config from '../../../lib/config';
 import { getBreakpoint } from '../../../lib/detect';
 import fastdom from '../../../lib/fastdom-promise';
 import { spaceFiller } from '../../common/modules/article/space-filler';
+import type {
+	SpacefinderRules,
+	SpacefinderWriter,
+} from '../../common/modules/article/space-filler';
 import { commercialFeatures } from '../../common/modules/commercial/commercial-features';
 import { addSlot } from './dfp/add-slot';
 import { createAdSlot } from './dfp/create-slot';
