+++ static/src/javascripts/projects/commercial/types.d.ts
@@ -29,42 +29,3 @@ type RuleSpacing = {
 	minAbove: number;
 	minBelow: number;
 };
-
-type SpacefinderItem = {
-	top: number;
-	bottom: number;
-	element: HTMLElement;
-};
-
-type SpacefinderRules = {
-	bodySelector: string;
-	body?: Node;
-	slotSelector: string;
-	// minimum from slot to top of page
-	absoluteMinAbove?: number;
-	// minimum from para to top of article
-	minAbove: number;
-	// minimum from (top of) para to bottom of article
-	minBelow: number;
-	// vertical px to clear the content meta element (byline etc) by. 0 to ignore
-	// used for carrot ads
-	clearContentMeta?: number;
-	// custom rules using selectors.
-	selectors: Record<string, RuleSpacing>;
-	// will run each slot through this fn to check if it must be counted in
-	filter?: (x: SpacefinderItem) => boolean;
-	// will remove slots before this one
-	startAt?: HTMLElement;
-	// will remove slots from this one on
-	stopAt?: HTMLElement;
-	// will reverse the order of slots (this is useful for lazy loaded content)
-	fromBottom?: boolean;
-};
-
-type SpacefinderWriter = (paras: HTMLElement[]) => void;
-
-type SpacefinderOptions = {
-	waitForLinks?: boolean;
-	waitForImages?: boolean;
-	waitForInteractives?: boolean;
-};
