+++ static/src/javascripts/projects/common/modules/article/space-filler.ts
@@ -2,6 +2,46 @@ import { findSpace, SpaceError } from 'common/modules/spacefinder';
 import fastdom from 'lib/fastdom-promise';
 import raven from 'lib/raven';
 
+// TODO move these types to spacefinder once it is converted to typescript
+type SpacefinderItem = {
+	top: number;
+	bottom: number;
+	element: HTMLElement;
+};
+
+type SpacefinderRules = {
+	bodySelector: string;
+	body?: Node;
+	slotSelector: string;
+	// minimum from slot to top of page
+	absoluteMinAbove?: number;
+	// minimum from para to top of article
+	minAbove: number;
+	// minimum from (top of) para to bottom of article
+	minBelow: number;
+	// vertical px to clear the content meta element (byline etc) by. 0 to ignore
+	// used for carrot ads
+	clearContentMeta?: number;
+	// custom rules using selectors.
+	selectors: Record<string, RuleSpacing>;
+	// will run each slot through this fn to check if it must be counted in
+	filter?: (x: SpacefinderItem) => boolean;
+	// will remove slots before this one
+	startAt?: HTMLElement;
+	// will remove slots from this one on
+	stopAt?: HTMLElement;
+	// will reverse the order of slots (this is useful for lazy loaded content)
+	fromBottom?: boolean;
+};
+
+type SpacefinderWriter = (paras: HTMLElement[]) => void;
+
+type SpacefinderOptions = {
+	waitForLinks?: boolean;
+	waitForImages?: boolean;
+	waitForInteractives?: boolean;
+};
+
 class SpaceFiller {
 	queue = Promise.resolve(true);
 
@@ -43,4 +83,4 @@ class SpaceFiller {
 
 const spaceFiller = new SpaceFiller();
 
-export { spaceFiller };
+export { spaceFiller, SpacefinderItem, SpacefinderRules, SpacefinderWriter };
